Details factored out as extensions.
