+++ src/DynamoWebServer/Messages/MessageHandler.cs
@@ -494,60 +494,22 @@ namespace DynamoWebServer.Messages
 
         private string GetInOutPortsData(NodeModel node)
         {
-            var pattern = node is VariableInputNode ? "{{ \"name\":\"{0}\", \"type\": \"{1}\" }}" : "\"{0}\"";
-            var inPorts = node.InPorts.Select(port => string.Format(pattern, port.PortName, port.ToolTipContent)).ToList();
-            var outPorts = node.OutPorts.Select(port => "\"" + port.ToolTipContent + "\"").ToList();
+            var isVarInputNode = false;
+            var stringBuilder = new StringBuilder("{");
 
-            var stringBuilder = new StringBuilder();
-
-            stringBuilder.Append("{");
             if (node is CodeBlockNodeModel)
             {
-                var codeBlock = node as CodeBlockNodeModel;
-                var allDefs = CodeBlockUtils.GetDefinitionLineIndexMap(codeBlock.CodeStatements);
-                var lineIndices = new List<int>();
-
-                foreach (var def in allDefs)
-                {
-                    var logicalIndex = def.Value - 1;
-                    lineIndices.Add(logicalIndex);
-                }
-
-                stringBuilder.Append("\"Code\":\"");
-                stringBuilder.Append(codeBlock.Code.
-                                     Replace("\n", "\\n").
-                                     Replace("\"", "\\\""));
-                stringBuilder.Append("\", ");
-                stringBuilder.Append("\"LineIndices\": [");
-                stringBuilder.Append(string.Join(", ", lineIndices.Select(x => x.ToString()).ToArray()));
-                stringBuilder.Append("],");
-            }
-
-            if (node is VariableInputNode)
+                stringBuilder.Append((node as CodeBlockNodeModel).GetExtraData());
+            } 
+            else if (node is VariableInputNode)
             {
-                var type = node.GetType();
-                if (type.Name == "PythonNode")
-                {
-                    var script = type.GetProperty("Script").GetValue(node, null).ToString();
-
-                    stringBuilder.Append("\"Script\":\"");
-                    stringBuilder.Append(script.
-                            Replace("\n", "\\n").
-                            Replace("\"", "\\\""));
-                    stringBuilder.Append("\", ");
-                }
-
-                stringBuilder.Append("\"varInputs\": [");
-            }
-            else
-            {
-                stringBuilder.Append("\"InPorts\": [");
+                stringBuilder.Append((node as VariableInputNode).GetExtraData());
+                isVarInputNode = true;
             }
 
-            stringBuilder.Append(inPorts.Any() ? inPorts.Aggregate((i, j) => i + "," + j) : "");
-            stringBuilder.Append("], \"OutPorts\": [");
-            stringBuilder.Append(outPorts.Any() ? outPorts.Aggregate((i, j) => i + "," + j) : "");
-            stringBuilder.Append("], \"Data\": \"" + GetValue(node) + "\"}");
+            stringBuilder.Append(node.GetInOutPortsData(isVarInputNode));
+            stringBuilder.Append(", \"Data\": \"" + GetValue(node));
+            stringBuilder.Append("\"}");
 
             return stringBuilder.ToString();
         }
